requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
fastapi
transformers
pillow
httpx
pytest
scikit-learn